requirements.txt
fastapi
uvicorn
python-dotenv
pdf2image
pillow
python-pptx
supabase

